chore: refactor triggers and actions to separate files
