+++ constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ZAPIER_TAG: 'created_by_zapier',
+};
